+++ Integrator.h
@@ -4,12 +4,12 @@
 #include "State.h"
 #include <functional>
 
-using DerivativeFn = std::function<State(float t, const State& s)>;
+using DerivativeFn = std::function<State(double t, const State& s)>;
 
 class Integrator {
 public:
     virtual ~Integrator() = default;
-    virtual State step(const State& state, DerivativeFn fn, float t, float dt) = 0;
+    virtual State step(const State& state, DerivativeFn fn, double t, double dt) = 0;
 };
 
 #endif
